coral-app/requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
brotlipy==0.7.0
cchardet==2.1.7
cffi==1.14.6
chardet==4.0.0
ffmpeg-python==0.2.0
future==0.18.2
idna==3.2
imageio==2.9.0
multidict==5.1.0
numpy==1.21.0
opencv-contrib-python==4.5.3.56
Pillow==8.3.1
pycares==4.0.0
pycparser==2.20
pyzmq==22.1.0
typing-extensions==3.10.0.0
yarl==1.6.3
